why should we use useMemo?
In this example here, we defining an array of 29 million items,
and searching it through is very expensive.

Every time the UI is rerendered, the search is started again and again
we need a caching mechanism for this so we don't use unnecessary computing power

if `items` is the same between renders, we don't need to recompute this at all.

Signed-off-by: John Titor
<50095635+JohnRTitor@users.noreply.github.com>
Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>