use useMemo to cache the expensive computation result
This will solve the performance issue of our component as the
code inside the useMemo won't be recomputed /rerun until any of
the dependencies in the array change.
Make sure to include all the dependencies in the dependency array.

Note that: we should only use useMemo only when necessary, it's because
it compares the values of the dependencies behind the scene so that might
cause performance issues (but not much). Just don't overuse it

Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>